requirements.txt
alembic==1.8.1
Babel==2.10.3
beautifulsoup4==4.11.1
blinker==1.5
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
elastic-transport==8.4.0
elasticsearch==8.4.1
email-validator==1.2.1
filelock==3.8.0
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
guess-language-spirit==0.5.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.4.4
protobuf==3.20.1
PyJWT==2.4.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
tweepy==4.10.1
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.1
WTForms==3.0.1
zipp==3.8.1
